Updated MongoURI to local and added async/await for seed
The async/await on the seed main function allows a much larger number of records to be created at once by leveraging the async/await method of allowing the event thread to proceed while not over flowing the memory heap tyring to keep everything in mind. Async/await, instead, uses pointers so that it can reconstruct the stack trace if necessary.