Mongo successfully seeded with 10m records; slower than PG
Though we use bulkWrite to insert batches of 10,000, it appears to be an order of magnitude slower than Postgres.
To insert 10m documents takes ~1hr.

Still, this solution does *not* rely on writing to/reading from disk, which in a space constrained environment like the cloud will be considered a victory.