Updated App.jsx to allow for local/remote hosting
Previously, the host was hardcoded in the baseURL as a localhost. This was sufficient when the service was hosted locally. With a move toward hosting remotely, however, this needed to be updated and will now follow the same environment as the Database, though the host is specified separately from the DB.